requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
greenlet==3.0.3
h11==0.14.0
icalendar==5.0.12
idna==3.7
outcome==1.3.0.post0
psycopg2-binary==2.9.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
schedule==1.2.2
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
wsproto==1.2.0
